# webui/web/openapi.json
{
  "components": {
    "schemas": {
...
      "GenerateAllRequest": {
        "properties": {
          "actions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actions"
          },
          "auto_accept": {
            "default": true,
            "title": "Auto Accept",
            "type": "boolean"
          },
          "max_regenerations": {
            "default": 1,
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Max Regenerations",
            "type": "integer"
          }
        },
        "title": "GenerateAllRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IdentityAnalyzeRequest": {
        "properties": {
          "force": {
            "default": false,
            "title": "Force",
            "type": "boolean"
          }
        },
        "title": "IdentityAnalyzeRequest",
        "type": "object"
      },
...
      "Job": {
        "properties": {
          "action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          },
          "attempt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempt"
          },
          "character": {
            "title": "Character",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "direction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction"
          },
          "elapsed_s": {
            "default": 0,
            "title": "Elapsed S",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobError"
              },
              {
                "type": "null"
              }
            ]
          },
          "expected_s": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected S"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "log": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/LogLine"
            },
            "title": "Log",
            "type": "array"
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Progress"
              },
              {
                "type": "null"
              }
            ]
          },
          "queue_position": {
            "default": 0,
            "title": "Queue Position",
            "type": "integer"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "default": "queued",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "canceled",
              "interrupted"
            ],
            "title": "State",
            "type": "string"
          },
          "type": {
            "enum": [
              "reference_generate",
              "identity_analyze",
              "action_generate",
              "batch_generate",
              "vision_review"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "character",
          "created_at"
        ],
        "title": "Job",
        "type": "object"
      },
      "JobCreated": {
        "properties": {
          "attempt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempt"
          },
          "job": {
            "$ref": "#/components/schemas/Job"
          }
        },
        "required": [
          "job"
        ],
        "title": "JobCreated",
        "type": "object"
      },
      "JobError": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "additionalProperties": true,
            "default": {},
            "title": "Detail",
            "type": "object"
          },
          "message": {
            "default": "",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "JobError",
        "type": "object"
      },
...
      "LogLine": {
        "properties": {
          "level": {
            "title": "Level",
            "type": "string"
          },
          "t": {
            "title": "T",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "t",
          "level",
          "text"
        ],
        "title": "LogLine",
        "type": "object"
      },
...
      "Progress": {
        "properties": {
          "current": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current"
          },
          "done": {
            "title": "Done",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "done",
          "total"
        ],
        "title": "Progress",
        "type": "object"
      },
...
      "ReferenceGenerateRequest": {
        "properties": {
          "count": {
            "default": 1,
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Count",
            "type": "integer"
          },
          "description": {
            "minLength": 1,
            "title": "Description",
            "type": "string"
          }
        },
        "required": [
          "description"
        ],
        "title": "ReferenceGenerateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "sprite-forge-web",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/characters/{cid}/generate-all": {
      "post": {
        "operationId": "generate_all_api_characters__cid__generate_all_post",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/GenerateAllRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate All"
      }
    },
...
    "/api/characters/{cid}/identity/analyze": {
      "post": {
        "operationId": "analyze_identity_api_characters__cid__identity_analyze_post",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/IdentityAnalyzeRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Identity"
      }
    },
...
    "/api/characters/{cid}/reference/generate": {
      "post": {
        "operationId": "generate_reference_api_characters__cid__reference_generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReferenceGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Reference"
      }
    },
...
  }
}